requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
eth-account>=0.10.0
web3>=6.0.0

